backend/flask/requirements.txt
apscheduler
Flask
flask-cors
Flask-HTTPAuth
numpy
psycopg2-binary
requests
uWSGI
